+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ```
 
+
 ## 0.0.5
 
 Added `changeAlertType()` function to change current dialog type.